+++ build.gradle
@@ -222,7 +222,7 @@ def versions = [
         reformLogging   : '5.1.7',
         sonarPitest     : '0.5',
         springfoxSwagger: '2.8.0',
-        sscsCommon      : '4.7.1-RC-SSCS-10220-4'
+        sscsCommon      : '4.7.1-RC-SSCS-10220-5'
 ]
 
 ext["rest-assured.version"] = '4.2.0'
